+++ packages/vue/src/accordion/accordion-item.tsx
@@ -1,13 +1,12 @@
 import { computed, defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
 import type { Assign } from '../types'
-import { type ComponentWithProps } from '../utils'
 import { useAccordionContext } from './accordion-context'
 import { AccordionItemProvider, type AccordionItemContext } from './accordion-item-context'
 
 export type AccordionItemProps = Assign<AccordionItemContext, HTMLArkProps<'div'>>
 
-export const AccordionItem: ComponentWithProps<AccordionItemProps> = defineComponent({
+export const AccordionItem = defineComponent({
   name: 'AccordionItem',
   props: {
     value: {
@@ -21,8 +20,7 @@ export const AccordionItem: ComponentWithProps<AccordionItemProps> = defineCompo
   },
   setup(props, { slots, attrs }) {
     const api = useAccordionContext()
-
-    AccordionItemProvider({ ...props, ...api.value.getItemState(props) })
+    AccordionItemProvider(computed(() => ({ ...props, ...api.value.getItemState(props) })))
     const itemState = computed(() => api.value.getItemState(props))
 
     return () => (